Type conversion bug.
git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/trunk@5389 bbc10eb1-c90d-0410-af57-cb519fbb1720

+++ prism/src/parser/type/TypeDouble.java
@@ -75,7 +75,7 @@ public class TypeDouble extends Type
 		if (value instanceof Double)
 			return (Double) value;
 		if (value instanceof Integer)
-			return new Double(((Double) value).doubleValue());
+			return new Double(((Integer) value).intValue());
 		else
 			throw new PrismLangException("Can't convert " + value.getClass() + " to type " + getTypeString());
 	}
